opensm/osm_ucast_mgr: better lft setup
The function set_next_lft_block() is called in loop with block number
incremented, inside it loops by itself in looking for changed block,
caller will call this function with original block number incremented
so this internal loop could be repeated again and again. This patch
cleans this ineffectiveness.

Also rename it to set_lft_block() since block number is treated as
parameters and *not* next block is processed and merges some code.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>