opensm: use mgrp pointer in port mcm_info
Port needs to access multicast groups where it is joined to. Now it is
implemented by keeping list of list of mcm_info elements where MLID of
each multicast group is stored. Obviously this assumes single MGID to
MLID mapping model.

This patch changes this so that instead of MLID mcm_info stores pointer
to multicast group object (mgrp). Such model makes it possible to
have MGIDs to MLID compression.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>