opensm SA DB: dump only if modified
Currently, if dumping SA DB is enabled, the file is dumped
at every sweep. This patch adds "dirty" flag that denotes
whether the SA DB needs to be dumped or not.

SA DB contains 3 types of data:
 - informinfo records
 - service records
 - mcmember records

So the flag must be set in 6 places:
 - add/remove informinfo record
 - add/remove service records
 - add/remove mcmember records

Another possible place for setting this flag is mcgroup
create/delete (this is sometimes duplicated by the add/remove
mcmember records), but I think that this is not needed:

The mcgroups w/o mcmembers can be created in two ways:

1. By creating well-known mcgroup for a partition.
   In this case when new instance of OSM is launched,
   the same mcgroup should be creted anyway from the
   partition configuration, so no need to force its
   duplication in the SA DB file.

2. By loading mcgroup from SA DB file.
   In this case it's either mcgroup w/o members, which
   is what's handled by (1), or it's mcgroup with members,
   butthe members just weren't loaded yet.

So in both cases, no need to force SA DB dump.

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>