Wrong handling of MC create and delete traps
For these traps the GID in the data details is the MGID and
not the source port gid.
So the SM should check that subscriber port has the pkey of the MC group.
There was also an error in comparing the subnet prefix and guid due to
host/network order mismatch.

Signed-off-by: Eli Dorfman <elid@voltaire.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>