opensm/osm_mcast_mgr.c: fix bug in MC root switch calculation
In newly introduced MC group root switch calculation method as it is
now, sw->num_of_mcm field (number of MC group members connected to this
switch) is updated only once when switch is first visited (for some
reason it is under such 'if' block), as result it will always '1' or
'0', which is obviously not correct and breaks the following hops
calculations algorithm. Fix this - update sw->num_of_mcm field
unconditionally.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>