opensm: Add option to specify prefix to syslog messages
Added the option 'log_prefix' that allows users to specify a prefix to
syslog messages from OpenSM.
For example, in a dual-fabric (or dual-rail) IB cluster, the prefix for
the first fabric could be "mpi" and the other fabric could be "storage".
This helps users to identify the subnet in the syslog messages when
two or more instances of OpenSM run in a single node to manage multiple
fabrics.

Summary of changes:
 o Added the option 'log_prefix' to OpenSM command line and config file.
 o Enhanced OpenSM logging facility to include the specified prefix in
   syslog messages.

Signed-off-by: Arputham Benjamin <abenjamin@sgi.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>