opensm: Better handling of non responsive SMAs
In order to better handle non responsive SMAs (when link is physically up
but the SMA does not respond), a timeout based mechanism for SMPs is added
to better enable forward progress in a more timely fashion. So rather than
waiting for timeouts and outstanding wire SMPs to drop below some configured
value, there is also an additional limit introduced for transaction based SMPs.
These timeout based SMPs are capped at a configured maximum value.

Two new options are added for this:
max_wire_smps2 indicates the second (higher) limit of timeout based
SMPs supported. When this limit is reached, timeout based SMPs
are not longer sent (until the number of outstanding ones drops below
this limit).
max_smps_timeout indicates the number of microseconds used for the
timeout in between sending SMPs when the oustanding wire count is above
max_wire_smps and below max_wire_smps2. It defaults to the transaction
timeout times the transaction retries.

The timeout based SMP mechanism can be disabled by setting max_wire_smps2
to the same value as max_wire_smps. This is equivalent to the (current)
algorithm prior to this change. By default, this mechanism is disabled.

Signed-off-by: Hal Rosenstock <hal.rosenstock@gmail.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>