opensm/osm_trap_rcv.c: No need for heavy sweep when just NodeDescription changes
For trap 144s which indicate NodeDescription changed, there is no need for
a heavy sweep if that is the only change indicated by the trap. Determine that
by checking that there are no other change flags set and that the
capability mask is unchanged. Still sweep if this is not the case (there
is some other change).

Signed-off-by: Hal Rosenstock <hal.rosenstock@gmail.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>