osmtest/SA client: Only set attribute offset for RMPP operations
Most of what the osmtest SA client is doing is not RMPP.
For example, any SA GetTable request is not an RMPP operation
but is requesting an RMPP response.

SA AttributeOffset only needs to be set in an RMPP DATA packet
so this is not needed on the requesting side.

Signed-off-by: Hal Rosenstock <hal.rosenstock@gmail.com>
Signed-off-by: Alex Netes <alexne@mellanox.com>