opensm/osm_ucast_mgr.c: ensure osm_ucast_mgr_process() returns failure when no routing engine runs.
If all configured routing engines fail to initialize, the routing engine
list will be empty.  If opensm is also configured for no fallback routing
engine, then osm_ucast_mgr_process() can incorrectly return success even
though no routing engine will run.  The result of that is that heavy
sweeps are attempted in a tight loop, spamming opensm logs and the fabric.

With this fix, heavy sweeps occur only after a trap or the sweep interval,
reducing log spam and making it easier to find the reason for the failure.

Signed-off-by: Jim Schutt <jaschut@sandia.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>