OpenSM - The DnUp routing algorithm.
This routing algorithm operates in a very similar fashion to UpDn, but
is modified to allow optimal routing on certain network structures in
which uplinks and CA nodes are connected to the same switch nodes. (For
example Chinook at EMSL and RoadRunner at LANL.) In these networks the
optimal paths between nodes connected to a single chassis would remain
within the chassis.  However due to the uplinks being connected at the
same level of the network as the CA nodes UpDn will not allow these paths
to be used for communication between the CA nodes.

DnUp follows the same procedure as UpDn with a few differences.  Ranking
is based solely on the relative distance from CA nodes, any switch node
with a CA node directly attached is assigned a rank of 0 any switch
node without a CA node attached is assigned a rank of one greater than
the minimum rank of their neighbors. Transitions are also reversed;
The initial direction is down and only one transition to up
is allowed.  There is also an option which relaxes this restriction to
allow communication with switches nodes similar to the functionality of
connect_roots in UpDn.

Signed-off-by: Alex Netes <alexne@mellanox.com>