opensm/osm_pkey_mgr.c: In pkey_mgr_update_peer_port, better last block handling
PKey table capacities are not required to be multiples of the PKey table block
size (32 entries of 16 pkeys).

Current code could enable partition enforcement on the peer switch port
even if the last partition table block were truncated. In this case, it's
better to disable partition enforcement on those ports.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>
Signed-off-by: Alex Netes <alexne@mellanox.com>