opensm/osm_sa_path_record.c: Add support for alias GUIDs
Use alias_port_guid_tbl to resolve SA path queries.
Any alias GUIDs in addition to the base port GUID
are present in the table based on SA sets and
deletes of GUIDInfoRecord.

Use this table for any DGID/SGID queries so if alias GUID is
used the alias GUID paths are also determined.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>
Signed-off-by: Alex Netes <alexne@mellanox.com>