opensm/osm_pkey_mgr.c: Pack switch peer port PKey tables
In pkey_mgr_update_peer_port, rather than mirror the end port
PKey table, pack the peer switch's port PKey table to
eliminate any holes.

This is only done for switch external ports and not end ports
because it's disruptive to change end port pkey indices
whereas that's not the case for switch external ports and it has benefit there.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>
Signed-off-by: Alex Netes <alexne@mellanox.com>