opensm: Add the precreation of multicast groups
Allow for the pre-creation of these groups on a partition by partition basis.

P_Key is taken from Partition Definition.  Q_Key, TClass, rate, FlowLabel and
mtu can be specified.

For IP groups, rate and mtu are verified to match the broadcast groups
parameters.  The P_Key bits in the mgid itself is verified to match the P_Key
in the Partition Definition.  If the P_Key bits in the mgid are 0x0000 then the
P_Key is taken from the Partition Definition.

The syntax extends the existing syntax by allowing MC groups to be specified
one per line, intermixed with the port specifications.

Changes from V1:

        Incorporate Hal's comments:
                Add FlowLabel
                copy documentation from man page to doc/partition-config.txt
                code format cleanup
                Remove "ERR:" from messages

        Move group flag parsing to common function for broadcast and generic
        groups.  Update documentation for the support of the other flags in the
        broadcast group.
        Error message clean up
        Update commit message

Signed-off-by: Ira Weiny <weiny2@llnl.gov>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>
Signed-off-by: Alex Netes <alexne@mellanox.com>