Optimized and deadlock-free routing algorithm for InfiniBand
Dear OpenSM community,

We would like to point you towards an optimized and
deadlock-free routing algorithm for InfiniBand. We finished
an implementation within OpenSM as part of an academic project.

The new routing resulted in up to 40% improved network performance
in some special architectures and it nearly halved the runtime
for all-to-all MPI communications on a 726-node production system,
located at the Technical University of Dresden.
Scientific paper describing the algorithm, implementation and
deadlock handling have been published at Hot Interconnects 2009 [2]
and IPDPS 2011 [1] repectively. You can find copies of all papers
at the bottom of [3].

We also provide a patch for the OpenSM integration, see below.
The patch, based on the latest checkout from Jan. 22, 2012, includes
two algorithms. The first one is the SSSP routing, which is not deadlock-free
but could used as an replacement for MinHop because it can also handle
arbitrary topologies. The second algorithm is DFSSSP, a deadlock-free
version of the SSSP routing, using virtual channels like the LASH algorithm
to avoid deadlocks. After a short review by Alex Netes, we updated the
documentation and the patch corresponds now to the code style of the
subnet manager.

We developed the algorithms in the course of an academic project.
Therefore our resources to maintain the patch are limited, however we
will answer any questions regarding the algorithm or implementation.
However, we believe that it is mostly bug-free and should be easy to
maintain by the community. We would appreciate if our work would find
its way into OpenSM (the license is BSD or GPLv2, like the other parts
of the subnet manager).

Thank you very much,
Jens Domke and Torsten Hoefler

[1]: Domke, Hoefler, Nagel: Deadlock-Free Oblivious Routing for
Arbitrary Topologies
(http://www.unixer.de/publications/index.php?pub=124)
[2]: Hoefler, Schneider, Lumsdaine: Optimized Routing for Large-Scale
InfiniBand Networks (http://www.unixer.de/publications/index.php?pub=93)
[3]: http://www.unixer.de/research/dfsssp/index.php

Signed-off-by: Jens Domke <jens.domke@tu-dresden.de>
Signed-off-by: Alex Netes <alexne@mellanox.com>