opensm: Fix opensm handover/relinquish corner case
When in master state and performing light sweeps, openSM ignores
other subnet managers in master state and fails to perform
handovers or relinquish control.  Handovers and relinquishing
master is only performed after a heavy sweep.  This can result
in two subnet managers being in master state, seeing each other
SM in master state, but both choose to do nothing about it.

This patch initiates a heavy sweep when another master subnet
manager is found during a light sweep.  This is sufficient to
start a handover or relinquish scenario.

Signed-off-by: Albert Chu <chu11@llnl.gov>
Signed-off-by: Alex Netes <alexne@mellanox.com>