opensm/osm_state_mgr.c: Force subn->need_update when coming out of STANDBY
When SM comes out of Standby, it might have data saved from the
time it was previously MASTER. Because of that it might not send
LFTs, pkeys and QoS configuration when it becomes MASTER again as
it might see that there is no change in the configuration.
But because we don't know what previous SM really configured, we
must send LFTs, pkeys, and QoS configuration to all switches.

Signed-off-by: Daniel Klein <danielk@mellanox.com>
Signed-off-by: Vladimir Koushnir <vladimirk@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>
Signed-off-by: Alex Netes <alexne@mellanox.com>