opensm/scripts/sldd.sh: Update to support guid2mkey/neighbors
Essentially a straightforward parameterization of all of the
functions, with a for loop wrapped around the main body, so
that the script can support sync'ing multiple files.

This allows us to add the new guid2mkey and neighbors cache
files to the default file list.

Semantics for the CACHE_FILE environment parameter have changed
slightly; instead of specifying a path to the guid2lid file,
it specifies a colon-separated list of paths to be synchronized,
which replaces the default list in its entirety.

Signed-off-by: Jim Foraker <foraker1@llnl.gov>
Signed-off-by: Alex Netes <alexne@mellanox.com>