Add cleanup of SA cache after handover
Previously, when SM becomes STANDBY after being MASTER, it preserved
the SA cache. When the SM will become MASTER again, it's SA cache
might be inconsistent. The solution is to clean the SA cache each
time that the SM becomes STANDBY after a handover.

Also, enhance drop_event_subscriptions option for ServiceRecords

drop_event_subscriptions is supposed to be just drop SA event subscriptions
(InformInfos). AGUID and MC subscriptions are always dropped
as client reregister mechanism works there.

As we are aware of potential issues with ServiceRecords (not being
reregistered), dropping of those is also controlled by this option.

Signed-off-by: Alex Netes <alexne@mellanox.com>
Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>